fix: failed to create python3 venv on Windows
